+++ pennylane_lightning/core/src/observables/Observables.hpp
@@ -62,6 +62,22 @@ template <class StateVectorT> class Observable {
     virtual void applyInPlace(StateVectorT &sv) const = 0;
 
     /**
+     * @brief Apply unitaries of an observable to the given statevector in
+     * place.
+     *
+     * @param sv Reference to StateVector object.
+     * @param identity_wires Reference to a std::vector object which stores
+     * wires of Identity gates in the observable.
+     * @param ob_wires Reference to a std::vector object which stores wires of
+     * the observable.
+     * @param term_idx Index of a Hamiltonian term.
+     */
+    virtual void applyInPlaceShots(StateVectorT &sv,
+                                   std::vector<size_t> &identity_wires,
+                                   std::vector<size_t> &ob_wires,
+                                   size_t term_idx = 0) const = 0;
+
+    /**
      * @brief Get the name of the observable
      */
     [[nodiscard]] virtual auto getObsName() const -> std::string = 0;
@@ -70,6 +86,13 @@ template <class StateVectorT> class Observable {
      * @brief Get the wires the observable applies to.
      */
     [[nodiscard]] virtual auto getWires() const -> std::vector<size_t> = 0;
+
+    /**
+     * @brief Get the coefficients of a Hamiltonian observable.
+     */
+    [[nodiscard]] virtual auto getCoeffs() const -> std::vector<PrecisionT> {
+        return {};
+    };
 
     /**
      * @brief Test whether this object is equal to another object
@@ -140,6 +163,32 @@ class NamedObsBase : public Observable<StateVectorT> {
     void applyInPlace(StateVectorT &sv) const override {
         sv.applyOperation(obs_name_, wires_, false, params_);
     }
+
+    void
+    applyInPlaceShots(StateVectorT &sv, std::vector<size_t> &identity_wire,
+                      std::vector<size_t> &ob_wires,
+                      [[maybe_unused]] size_t term_idx = 0) const override {
+        ob_wires.clear();
+        identity_wire.clear();
+        ob_wires.push_back(wires_[0]);
+
+        if (obs_name_ == "PauliX") {
+            sv.applyOperation("Hadamard", wires_, false);
+        } else if (obs_name_ == "PauliY") {
+            sv.applyOperations({"PauliZ", "S", "Hadamard"},
+                               {wires_, wires_, wires_}, {false, false, false});
+        } else if (obs_name_ == "Hadamard") {
+            const PrecisionT theta = -M_PI / 4.0;
+            sv.applyOperation("RY", wires_, false, {theta});
+        } else if (obs_name_ == "PauliZ") {
+        } else if (obs_name_ == "Identity") {
+            identity_wire.push_back(wires_[0]);
+        } else {
+            PL_ABORT("Provided NamedObs does not supported for shots "
+                     "calculation. Supported NamedObs are PauliX, PauliY, "
+                     "PauliZ, Identity and Hadamard.");
+        }
+    }
 };
 
 /**
@@ -192,6 +241,15 @@ class HermitianObsBase : public Observable<StateVectorT> {
     void applyInPlace(StateVectorT &sv) const override {
         sv.applyMatrix(matrix_, wires_);
     }
+
+    void
+    applyInPlaceShots([[maybe_unused]] StateVectorT &sv,
+                      [[maybe_unused]] std::vector<size_t> &identity_wire,
+                      [[maybe_unused]] std::vector<size_t> &ob_wires,
+                      [[maybe_unused]] size_t term_idx = 0) const override {
+        PL_ABORT("Hermitian observables do not support applyInPlaceShots "
+                 "method.");
+    }
 };
 
 /**
@@ -224,6 +282,7 @@ class TensorProdObsBase : public Observable<StateVectorT> {
     }
 
   public:
+    using PrecisionT = typename StateVectorT::PrecisionT;
     /**
      * @brief Create a tensor product of observables
      *
@@ -298,6 +357,23 @@ class TensorProdObsBase : public Observable<StateVectorT> {
     void applyInPlace(StateVectorT &sv) const override {
         for (const auto &ob : obs_) {
             ob->applyInPlace(sv);
+        }
+    }
+
+    void
+    applyInPlaceShots(StateVectorT &sv, std::vector<size_t> &identity_wires,
+                      std::vector<size_t> &ob_wires,
+                      [[maybe_unused]] size_t term_idx = 0) const override {
+        identity_wires.clear();
+        ob_wires.clear();
+        for (const auto &ob : obs_) {
+            std::vector<size_t> identity_wire;
+            std::vector<size_t> ob_wire;
+            ob->applyInPlaceShots(sv, identity_wire, ob_wire);
+            if (!identity_wire.empty()) {
+                identity_wires.push_back(identity_wire[0]);
+            }
+            ob_wires.push_back(ob_wire[0]);
         }
     }
 
@@ -386,6 +462,15 @@ class HamiltonianBase : public Observable<StateVectorT> {
                  "defined at the backend level.");
     }
 
+    void
+    applyInPlaceShots([[maybe_unused]] StateVectorT &sv,
+                      [[maybe_unused]] std::vector<size_t> &identity_wires,
+                      [[maybe_unused]] std::vector<size_t> &ob_wires,
+                      [[maybe_unused]] size_t term_idx = 0) const override {
+        PL_ABORT("For Hamiltonian Observables, the applyInPlace method must be "
+                 "defined at the backend level.");
+    }
+
     [[nodiscard]] auto getWires() const -> std::vector<size_t> override {
         std::unordered_set<size_t> wires;
 
@@ -412,6 +497,13 @@ class HamiltonianBase : public Observable<StateVectorT> {
         ss << "]}";
         return ss.str();
     }
+
+    /**
+     * @brief Get the coefficients of the observable.
+     */
+    [[nodiscard]] auto getCoeffs() const -> std::vector<PrecisionT> override {
+        return coeffs_;
+    };
 };
 
 /**
@@ -494,6 +586,15 @@ class SparseHamiltonianBase : public Observable<StateVectorT> {
         PL_ABORT("For SparseHamiltonian Observables, the applyInPlace method "
                  "must be "
                  "defined at the backend level.");
+    }
+
+    void
+    applyInPlaceShots([[maybe_unused]] StateVectorT &sv,
+                      [[maybe_unused]] std::vector<size_t> &identity_wire,
+                      [[maybe_unused]] std::vector<size_t> &ob_wires,
+                      [[maybe_unused]] size_t term_idx = 0) const override {
+        PL_ABORT("SparseHamiltonian observables do not the applyInPlaceShots "
+                 "method.");
     }
 
     [[nodiscard]] auto getObsName() const -> std::string override {