+++ pennylane_lightning/core/src/simulators/lightning_qubit/observables/ObservablesLQubit.hpp
@@ -358,6 +358,16 @@ class Hamiltonian final : public HamiltonianBase<StateVectorT> {
             StateVectorT, Pennylane::Util::use_openmp>::run(this->coeffs_,
                                                             this->obs_, sv);
     }
+
+    // to work with
+    void applyInPlaceShots(StateVectorT &sv,
+                           std::vector<size_t> &identity_wires,
+                           std::vector<size_t> &ob_wires,
+                           size_t term_idx) const override {
+        ob_wires.clear();
+        this->obs_[term_idx]->applyInPlaceShots(sv, identity_wires, ob_wires,
+                                                term_idx);
+    }
 };
 
 /**